Use the object_file.lst of the modules
The prior method was failing for some cases, where the obj files where
in a deeper directory, X64 (always?).

I found that this file have the path to the obj files so it's ok to
check there.

I already knew about these files, and I don't remember why I did drop
them and did all the thing with ar to dearchive the lib
files. Hopefully everything is good?

To check that it is good I grep the logs, comparing the times the
script is called vs the times the edition is calculated (grep
'edition=')