Use a single source_files.lst instead of *.obj.src
I started this change to be more consistent with the already used
object_files.lst and ended up finding out some mistakes in the string
manipulations in edition.sh plus that this method is using more files
to make the hash.

About that mistake; I wasn't interpreting well the newlines and spaces
in the strings causing sort to act badly. Plus now I'm filtering by
uniqueness while sorting, just for prevention, since now I'm appending
files to the *.lst instead of sending them altogheter.

Plus I did quite a lot of debugging so I'm confident that this works
better.

Just to document, for debug I logged or the hashes with the module
name so I can do a quick diff between changes to see what's going on.

Something that was quite useful was

    find Build -name '*.files' -exec sh -c 'cp $0 shell-files-good/$(basename $0)' {} \;

to save the *.files for comparission between versions, plus looking at
the build logs and hashes logs.